+++ tests/unit/test_elf.py
@@ -20,7 +20,7 @@ import tempfile
 from textwrap import dedent
 import sys
 
-from testtools.matchers import Contains, Equals, NotEquals
+from testtools.matchers import Contains, EndsWith, Equals, NotEquals
 from unittest import mock
 
 from snapcraft.internal import errors, elf, os_release
@@ -472,3 +472,52 @@ class TestSonameCache(unit.TestCase):
         self.assertFalse('soname.so' in self.soname_cache)
         self.assertFalse('notfound.so' in self.soname_cache)
         self.assertTrue('soname2.so' in self.soname_cache)
+
+
+# This is just a subset
+_LIBC6_LIBRARIES = [
+    'ld-2.26.so',
+    'ld-linux-x86-64.so.2',
+    'libBrokenLocale-2.26.so',
+    'libBrokenLocale.so.1',
+    'libSegFault.so',
+    'libanl-2.26.so',
+]
+
+
+class HandleGlibcTestCase(unit.TestCase):
+
+    def _setup_libc6(self):
+        lib_path = os.path.join(self.path, 'lib')
+        libraries = {os.path.join(lib_path, l) for l in _LIBC6_LIBRARIES}
+
+        os.mkdir(lib_path)
+        for library in libraries:
+            open(library, 'w').close()
+
+        return libraries
+
+    def setUp(self):
+        super().setUp()
+
+        patcher = mock.patch(
+            'snapcraft.internal.repo.Repo.get_package_libraries')
+        self.get_packages_mock = patcher.start()
+        self.get_packages_mock.return_value = self._setup_libc6()
+        self.addCleanup(patcher.stop)
+
+    def test_glibc_mangling(self):
+        dynamic_linker = elf.find_linker(
+            root_path=self.path,
+            snap_base_path='/snap/snap-name/current')
+
+        self.get_packages_mock.assert_called_once_with('libc6')
+
+        self.assertThat(dynamic_linker, EndsWith('ld-2.26.so'))
+
+    def test_bad_dynamic_linker_in_libc6_package(self):
+        self.get_packages_mock.return_value = {'/usr/lib/dyn-linker-2.25.so'}
+        self.assertRaises(RuntimeError,
+                          elf.find_linker,
+                          root_path=self.path,
+                          snap_base_path='/snap/snap-name/current')
