+++ tests/unit/test_mangling.py
@@ -16,9 +16,8 @@
 
 import os
 import textwrap
-from unittest.mock import patch
 
-from testtools.matchers import EndsWith, FileContains, FileExists, Not
+from testtools.matchers import FileContains, FileExists, Not
 
 from snapcraft.internal import mangling
 from tests import unit, fixture_setup
@@ -107,56 +106,6 @@ class ManglingPythonShebangTestCase(unit.TestCase):
             This is a test
             =======================
         """)))
-
-
-# This is just a subset
-_LIBC6_LIBRARIES = [
-    'ld-2.26.so',
-    'ld-linux-x86-64.so.2',
-    'libBrokenLocale-2.26.so',
-    'libBrokenLocale.so.1',
-    'libSegFault.so',
-    'libanl-2.26.so',
-]
-
-
-class HandleGlibcTestCase(unit.TestCase):
-
-    def _setup_libc6(self):
-        lib_path = os.path.join(self.path, 'lib')
-        libraries = {os.path.join(lib_path, l) for l in _LIBC6_LIBRARIES}
-
-        os.mkdir(lib_path)
-        for library in libraries:
-            open(library, 'w').close()
-
-        return libraries
-
-    def setUp(self):
-        super().setUp()
-
-        patcher = patch('snapcraft.internal.repo.Repo.get_package_libraries')
-        self.get_packages_mock = patcher.start()
-        self.get_packages_mock.return_value = self._setup_libc6()
-        self.addCleanup(patcher.stop)
-
-    def test_glibc_mangling(self):
-        dynamic_linker = mangling.handle_glibc_mismatch(
-            root_path=self.path,
-            core_base_path='/snap/core/current',
-            snap_base_path='/snap/snap-name/current')
-
-        self.get_packages_mock.assert_called_once_with('libc6')
-
-        self.assertThat(dynamic_linker, EndsWith('ld-2.26.so'))
-
-    def test_bad_dynamic_linker_in_libc6_package(self):
-        self.get_packages_mock.return_value = {'/usr/lib/dyn-linker-2.25.so'}
-        self.assertRaises(RuntimeError,
-                          mangling.handle_glibc_mismatch,
-                          root_path=self.path,
-                          core_base_path='/snap/core/current',
-                          snap_base_path='/snap/snap-name/current')
 
 
 class TestClearExecstack(unit.TestCase):
